+++ runtime/src/main/java/org/corfudb/runtime/exceptions/TransactionAbortedException.java
@@ -15,7 +15,7 @@ public class TransactionAbortedException extends RuntimeException {
     TxResolutionInfo txResolutionInfo;
 
     @Getter
-    Long conflictKey;
+    Integer conflictKey;
 
     @Getter
     Throwable cause;
@@ -28,13 +28,13 @@ public class TransactionAbortedException extends RuntimeException {
      */
     public TransactionAbortedException(
             TxResolutionInfo txResolutionInfo,
-            Long conflictKey, AbortCause abortCause) {
+            Integer conflictKey, AbortCause abortCause) {
         this(txResolutionInfo, conflictKey, abortCause, null);
     }
 
     public TransactionAbortedException(
             TxResolutionInfo txResolutionInfo,
-            Long conflictKey, AbortCause abortCause, Throwable cause) {
+            Integer conflictKey, AbortCause abortCause, Throwable cause) {
         super("TX ABORT "
                 + " | Snapshot Time = " + txResolutionInfo.getSnapshotTimestamp()
                 + " | Transaction ID = " + txResolutionInfo.getTXid()
